Support specifying turnover rate
This adds support for specifying the turnover rate and interval, which
is useful if you don't want your series to be replaced by new ones.

The `--turnover-rate` flag specifies turnover rate, as the denominator
of a fraction.  If set to zero, it disables turnover, otherwise, at
least, 1/turnoverRate series will be replaced each interval, plus at
most the rest of the series in the metric.  The `--turnover-interval`
flag specifies the interval at which this occurs.